6: Realtid
Functions get- and saveMsgs are working when receiving data from mongodb from the backend.
Formatted the date to an Intl.DateTimeFormat object when creating the chat messages.
When saving the chat messages an easier date will be displayed, this so it also can be easier to search for a partiular date from the collection in mongodb.